Change TOP structure
