+++ src/components/pages/index/top/BooksTop.tsx
@@ -1,8 +1,10 @@
 import React, { FC } from "react"
 import { join } from "lodash-es"
+import { Box, Grid, Typography } from "@mui/material"
 import { Author, Book } from "@/api/types"
-import { Box, Typography } from "@mui/material"
 import { StyledImage } from "./styles"
+
+const TOP_BOOKS_AMOUNT = 4
 
 export const BookItem: FC<{ book: Book; author: Author }> = ({ book, author }) => {
   const authorFullName = join([author.firstName, author.lastName], " ")
@@ -19,5 +21,23 @@ export const BookItem: FC<{ book: Book; author: Author }> = ({ book, author }) =
         </Typography>
       </Box>
     </Box>
+  )
+}
+
+export const BooksTop: FC<{ books: Book[] }> = ({ books }) => {
+  const booksTop = books.slice(0, TOP_BOOKS_AMOUNT)
+  const topBooksTitle = `Top ${TOP_BOOKS_AMOUNT} books`
+
+  return (
+    <Grid container>
+      <Grid item xs={12} mb={1}>
+        <Typography variant="h4">{topBooksTitle}</Typography>
+      </Grid>
+      {booksTop.map((book) => (
+        <Grid item xs={12} md={6} key={book.id}>
+          <BookItem author={book.author as Author} book={book} />
+        </Grid>
+      ))}
+    </Grid>
   )
 }
